Fix off-board coordinate mishandling
In the event handler section for SDL_MOUSEBUTTONDOWN, calls to the
screen2field_x() and screen2field_y() functions are used to
calculate offsets on the board relative to pixels at OFFSET_X and
OFFSET_Y.

Since these functions return '0' for a band of pixels on either
side of the top [all y == OFFSET_Y] and left [all x == OFFSET_X]
edges of the board, some off-board pixels can lead to calls to
turn() which the validation logic does not generate
INVALID_FIELD_VALUE for.

Add tests that bail early from the event handler to prevent this.